Added missing cast.
Former-commit-id: 844afef3817a1c28484c508c39d6063022a81a0a
Former-commit-id: 5a1b1459dffe8a9cc703d918d446f6d18d810517
Former-commit-id: 5e44fe7f00545221926886f9ce70665ca4869d50

+++ src/ImageProcessorCore/Numerics/BigRational.cs
@@ -138,7 +138,7 @@ namespace ImageProcessorCore
             Denominator = 1;
 
             double val = Math.Abs(value);
-            double df = Numerator / Denominator;
+            double df = Numerator / (double)Denominator;
             double epsilon = bestPrecision ? double.Epsilon : .000001;
 
             while (Math.Abs(df - val) > epsilon)
